refactoring & formatting changes
