+++ Team 3749 2023/src/main/java/frc/robot/commands/vision/AlignApriltag.java
@@ -1,12 +1,9 @@
-package frc.robot.commands.swerve;
+package frc.robot.commands.vision;
 
 import frc.robot.utils.Constants;
 import frc.robot.utils.SmartData;
 import frc.robot.utils.Constants.VisionConstants;
-import frc.robot.utils.Constants.VisionConstants.Node;
-
 import org.photonvision.targeting.PhotonTrackedTarget;
-
 import edu.wpi.first.apriltag.AprilTagFieldLayout;
 import edu.wpi.first.math.controller.ProfiledPIDController;
 import edu.wpi.first.math.geometry.Pose2d;
