+++ Team 3749 2023/src/main/java/frc/robot/utils/JoystickIO.java
@@ -11,9 +11,9 @@ import edu.wpi.first.wpilibj2.command.CommandBase;
 import edu.wpi.first.wpilibj2.command.Commands;
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
 import frc.robot.commands.arm.MoveArm;
-import frc.robot.commands.swerve.AlignApriltag;
 import frc.robot.commands.swerve.AlignHeading;
-import frc.robot.commands.swerve.AlignRetro;
+import frc.robot.commands.vision.AlignApriltag;
+import frc.robot.commands.vision.AlignRetro;
 import frc.robot.subsystems.arm.Arm;
 import frc.robot.subsystems.intake.ArmIntake;
 import frc.robot.subsystems.intake.SideIntake;
